FinalProject/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
bcrypt==3.2.0
python-dotenv
pymysql
python-jose
datetime
python-dotenv
python-multipart
